feat(6806): extract tcp transport adapter tests
