+++ crates/kamn-sdk/tests/tcp_transport_adapter.rs
@@ -1,632 +1,12 @@
-use kamn_core::{
-    service_auth_public_key_hex_from_private_key_hex, service_auth_sign_with_private_key_hex,
-};
-use kamn_sdk::{
-    signature_for_fields, AgentDid, SdkError, TcpSignedEnvelope, TcpTransportAdapter,
-    TcpTransportConfig,
-};
-use std::io::Write;
-use std::net::{Shutdown, TcpListener, TcpStream};
-use std::thread;
-use std::time::{Duration, Instant};
-
-const TEST_TCP_SIGNING_PRIVATE_KEY_HEX: &str =
-    "658c3528422eb527b4c108b8f6d1e5f629543c304ea49cf608c67794424291c4";
-const TEST_TCP_SIGNING_PRIVATE_KEY_HEX_ALT: &str =
-    "094cf4e1f3d974bbf3e72233e2c2937e8fdb094740e0f017e010aa47ac1201ac";
-
-fn did_unbound(value: &str) -> AgentDid {
-    match AgentDid::parse(format!("kamn:did:agent:{value}")) {
-        Ok(parsed) => parsed,
-        Err(error) => panic!("did parse failed: {error}"),
-    }
-}
-
-fn did(value: &str) -> AgentDid {
-    let signer_public_key = signer_public_key_hex_for_private_key(TEST_TCP_SIGNING_PRIVATE_KEY_HEX);
-    match AgentDid::with_public_key_hex_binding(value, signer_public_key.as_str()) {
-        Ok(bound) => bound,
-        Err(error) => panic!("bound did parse failed: {error}"),
-    }
-}
-
-fn free_addr() -> String {
-    let listener = match TcpListener::bind("127.0.0.1:0") {
-        Ok(bound) => bound,
-        Err(error) => panic!("failed to allocate free tcp address: {error}"),
-    };
-    let addr = match listener.local_addr() {
-        Ok(local) => local,
-        Err(error) => panic!("failed to read allocated local address: {error}"),
-    };
-    addr.to_string()
-}
-
-fn send_raw_payload(addr: &str, payload: &str) {
-    for _attempt in 0..40 {
-        match TcpStream::connect(addr) {
-            Ok(mut stream) => {
-                if let Err(error) = stream.write_all(payload.as_bytes()) {
-                    panic!("failed to write raw payload: {error}");
-                }
-                if let Err(error) = stream.flush() {
-                    panic!("failed to flush raw payload: {error}");
-                }
-                if let Err(error) = stream.shutdown(Shutdown::Write) {
-                    panic!("failed to shutdown raw payload stream: {error}");
-                }
-                return;
-            }
-            Err(_) => thread::sleep(Duration::from_millis(10)),
-        }
-    }
-    panic!("failed to connect raw payload sender to {addr}");
-}
-
-fn signer_public_key_hex() -> String {
-    match service_auth_public_key_hex_from_private_key_hex(TEST_TCP_SIGNING_PRIVATE_KEY_HEX) {
-        Ok(value) => value,
-        Err(error) => panic!("failed to derive tcp signer public key: {error}"),
-    }
-}
-
-fn signer_public_key_hex_for_private_key(private_key_hex: &str) -> String {
-    match service_auth_public_key_hex_from_private_key_hex(private_key_hex) {
-        Ok(value) => value,
-        Err(error) => {
-            panic!("failed to derive tcp signer public key for {private_key_hex}: {error}")
-        }
-    }
-}
-
-#[test]
-fn unit_tcp_envelope_roundtrip_is_cryptographic() {
-    let envelope = match TcpSignedEnvelope::new(
-        did("sender-1"),
-        did("listener-1"),
-        7,
-        "state:runtime-7",
-        "hello-runtime",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope build failed: {error}"),
-    };
-
-    let wire = envelope.to_wire_payload();
-    let parsed = match TcpSignedEnvelope::parse_wire_payload(&wire) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope parse failed: {error}"),
-    };
-
-    assert_eq!(parsed, envelope);
-}
-
-#[test]
-fn unit_tcp_envelope_rejects_duplicate_keys() {
-    let payload = "from=kamn:did:agent:sender-1\nfrom=kamn:did:agent:sender-1\nto=kamn:did:agent:listener-1\nnonce=1\nstate_hash=state:dup\nbody=dup\nsignature=sig:deterministic-v1:baseline-v1:kamn:did:agent:sender-1:1:state:dup:3\n";
-
-    assert_eq!(
-        TcpSignedEnvelope::parse_wire_payload(payload),
-        Err(SdkError::InvalidInput {
-            field: "wire_payload",
-            reason: "duplicate key: from",
-        })
-    );
-}
-
-#[test]
-fn regression_tcp_envelope_rejects_baseline_v1_deterministic_signature() {
-    // Regression: #5846
-    let from = did("sender-legacy");
-    let to = did("listener-legacy");
-    let signer_public_key = signer_public_key_hex();
-    let signature = signature_for_fields(from.as_str(), 2, "state:legacy", "legacy-payload");
-    let payload = format!(
-        "from={}\n\
-to={}\n\
-nonce=2\n\
-state_hash=state:legacy\n\
-body=legacy-payload\n\
-signer_public_key={signer_public_key}\n\
-signature={signature}\n",
-        from.as_str(),
-        to.as_str()
-    );
-
-    assert_eq!(
-        TcpSignedEnvelope::parse_wire_payload(payload.as_str()),
-        Err(SdkError::InvalidInput {
-            field: "signature",
-            reason: "failed cryptographic envelope verification",
-        })
-    );
-}
-
-#[test]
-fn regression_tcp_envelope_rejects_missing_did_key_binding_fingerprint() {
-    // Regression: #6299
-    let from = did_unbound("sender-unbound");
-    let signer_public_key = signer_public_key_hex_for_private_key(TEST_TCP_SIGNING_PRIVATE_KEY_HEX);
-    let signature = service_auth_sign_with_private_key_hex(
-        from.as_str(),
-        3,
-        "state:binding",
-        "binding-check",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    )
-    .expect("signature");
-
-    let envelope = TcpSignedEnvelope {
-        from,
-        to: did("listener-bound"),
-        nonce: 3,
-        state_hash: "state:binding".to_owned(),
-        body: "binding-check".to_owned(),
-        signer_public_key,
-        signature,
-    };
-
-    assert_eq!(
-        envelope.verify_integrity(),
-        Err(SdkError::InvalidInput {
-            field: "from",
-            reason: "must include key-binding fingerprint matching signer_public_key",
-        })
-    );
-}
-
-#[test]
-fn regression_tcp_envelope_rejects_mismatched_did_key_binding_fingerprint() {
-    // Regression: #6299
-    let signer_public_key = signer_public_key_hex_for_private_key(TEST_TCP_SIGNING_PRIVATE_KEY_HEX);
-    let mismatched_binding_public_key =
-        signer_public_key_hex_for_private_key(TEST_TCP_SIGNING_PRIVATE_KEY_HEX_ALT);
-    let from = AgentDid::with_public_key_hex_binding(
-        "sender-bound-mismatch",
-        mismatched_binding_public_key.as_str(),
-    )
-    .expect("from did");
-    let signature = service_auth_sign_with_private_key_hex(
-        from.as_str(),
-        4,
-        "state:binding-mismatch",
-        "binding-mismatch-check",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    )
-    .expect("signature");
-
-    let envelope = TcpSignedEnvelope {
-        from,
-        to: did("listener-bound"),
-        nonce: 4,
-        state_hash: "state:binding-mismatch".to_owned(),
-        body: "binding-mismatch-check".to_owned(),
-        signer_public_key,
-        signature,
-    };
-
-    assert_eq!(
-        envelope.verify_integrity(),
-        Err(SdkError::InvalidInput {
-            field: "from",
-            reason: "must include key-binding fingerprint matching signer_public_key",
-        })
-    );
-}
-
-#[test]
-fn integration_tcp_parse_rejects_unbound_sender_did_key_binding() {
-    // Integration: #6299
-    let from = did_unbound("sender-wire-unbound");
-    let to = did("listener-wire-bound");
-    let signer_public_key = signer_public_key_hex_for_private_key(TEST_TCP_SIGNING_PRIVATE_KEY_HEX);
-    let signature = service_auth_sign_with_private_key_hex(
-        from.as_str(),
-        5,
-        "state:wire-binding",
-        "wire-binding-check",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    )
-    .expect("signature");
-    let payload = format!(
-        "from={}\nto={}\nnonce=5\nstate_hash=state:wire-binding\nbody=wire-binding-check\nsigner_public_key={}\nsignature={}\n",
-        from.as_str(),
-        to.as_str(),
-        signer_public_key,
-        signature
-    );
-
-    assert_eq!(
-        TcpSignedEnvelope::parse_wire_payload(payload.as_str()),
-        Err(SdkError::InvalidInput {
-            field: "from",
-            reason: "must include key-binding fingerprint matching signer_public_key",
-        })
-    );
-}
-
-#[test]
-fn functional_tcp_adapter_relays_signed_envelope_between_two_processes() {
-    let addr = free_addr();
-
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let sender_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("sender config failed: {error}"),
-    };
-
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-    let sender_adapter = TcpTransportAdapter::new(sender_config);
-
-    let expected_envelope = match TcpSignedEnvelope::new(
-        did("sender-func"),
-        did("listener-func"),
-        11,
-        "state:functional",
-        "functional-envelope",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope build failed: {error}"),
-    };
-
-    let listener_thread = thread::spawn(move || listener_adapter.listen_once());
-    thread::sleep(Duration::from_millis(30));
-
-    if let Err(error) = sender_adapter.send(&expected_envelope) {
-        panic!("sender adapter failed to send envelope: {error}");
-    }
-
-    let received = match listener_thread.join() {
-        Ok(result) => match result {
-            Ok(value) => value,
-            Err(error) => panic!("listener adapter failed: {error}"),
-        },
-        Err(_) => panic!("listener thread panicked"),
-    };
-
-    assert_eq!(received.envelope, expected_envelope);
-    assert!(received.peer_addr.starts_with("127.0.0.1:"));
-}
-
-#[test]
-fn integration_tcp_adapter_rejects_oversized_wire_payload() {
-    let addr = free_addr();
-
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => match value.with_max_wire_bytes(16) {
-            Ok(config) => config,
-            Err(error) => panic!("listener max-wire config failed: {error}"),
-        },
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let sender_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("sender config failed: {error}"),
-    };
-
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-    let sender_adapter = TcpTransportAdapter::new(sender_config);
-
-    let large_body = "x".repeat(64);
-    let envelope = match TcpSignedEnvelope::new(
-        did("sender-large"),
-        did("listener-large"),
-        1,
-        "state:large",
-        large_body,
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope build failed: {error}"),
-    };
-
-    let listener_thread = thread::spawn(move || listener_adapter.listen_once());
-    thread::sleep(Duration::from_millis(30));
-
-    if let Err(error) = sender_adapter.send(&envelope) {
-        panic!("sender adapter failed: {error}");
-    }
-
-    let listener_result = match listener_thread.join() {
-        Ok(result) => result,
-        Err(_) => panic!("listener thread panicked"),
-    };
-
-    assert_eq!(
-        listener_result,
-        Err(SdkError::InvalidInput {
-            field: "wire_payload",
-            reason: "exceeds max wire bytes",
-        })
-    );
-}
-
-#[test]
-fn functional_tcp_adapter_reconnect_preserves_nonce_replay_guard_state() {
-    let addr = free_addr();
-
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let sender_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("sender config failed: {error}"),
-    };
-
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-    let sender_adapter = TcpTransportAdapter::new(sender_config);
-
-    let first = match TcpSignedEnvelope::new(
-        did("sender-reconnect"),
-        did("listener-reconnect"),
-        1,
-        "state:reconnect",
-        "first-connect",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("first envelope build failed: {error}"),
-    };
-    let second = match TcpSignedEnvelope::new(
-        did("sender-reconnect"),
-        did("listener-reconnect"),
-        2,
-        "state:reconnect",
-        "second-connect",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("second envelope build failed: {error}"),
-    };
-
-    let first_listener = listener_adapter.clone();
-    let first_thread = thread::spawn(move || first_listener.listen_once());
-    thread::sleep(Duration::from_millis(30));
-
-    if let Err(error) = sender_adapter.send(&first) {
-        panic!("first send failed: {error}");
-    }
-    let first_received = match first_thread.join() {
-        Ok(result) => match result {
-            Ok(value) => value,
-            Err(error) => panic!("first listen failed: {error}"),
-        },
-        Err(_) => panic!("first listener thread panicked"),
-    };
-    assert_eq!(first_received.envelope, first);
-
-    let second_listener = listener_adapter.clone();
-    let second_thread = thread::spawn(move || second_listener.listen_once());
-    thread::sleep(Duration::from_millis(30));
-
-    if let Err(error) = sender_adapter.send(&second) {
-        panic!("second send failed: {error}");
-    }
-    let second_received = match second_thread.join() {
-        Ok(result) => match result {
-            Ok(value) => value,
-            Err(error) => panic!("second listen failed: {error}"),
-        },
-        Err(_) => panic!("second listener thread panicked"),
-    };
-    assert_eq!(second_received.envelope, second);
-}
-
-#[test]
-fn integration_tcp_adapter_replay_nonce_is_rejected_across_reconnect() {
-    let addr = free_addr();
-
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let sender_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("sender config failed: {error}"),
-    };
-
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-    let sender_adapter = TcpTransportAdapter::new(sender_config);
-
-    let first = match TcpSignedEnvelope::new(
-        did("sender-replay"),
-        did("listener-replay"),
-        9,
-        "state:replay",
-        "nonce-9-initial",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("first envelope build failed: {error}"),
-    };
-    let replayed = match TcpSignedEnvelope::new(
-        did("sender-replay"),
-        did("listener-replay"),
-        9,
-        "state:replay",
-        "nonce-9-replayed",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("replayed envelope build failed: {error}"),
-    };
-
-    let first_listener = listener_adapter.clone();
-    let first_thread = thread::spawn(move || first_listener.listen_once());
-    thread::sleep(Duration::from_millis(30));
-    if let Err(error) = sender_adapter.send(&first) {
-        panic!("first send failed: {error}");
-    }
-    match first_thread.join() {
-        Ok(result) => {
-            if let Err(error) = result {
-                panic!("first listen failed: {error}");
-            }
-        }
-        Err(_) => panic!("first listener thread panicked"),
-    }
-
-    let replay_listener = listener_adapter.clone();
-    let replay_thread = thread::spawn(move || replay_listener.listen_once());
-    thread::sleep(Duration::from_millis(30));
-    if let Err(error) = sender_adapter.send(&replayed) {
-        panic!("replay send failed: {error}");
-    }
-    let replay_result = match replay_thread.join() {
-        Ok(result) => result,
-        Err(_) => panic!("replay listener thread panicked"),
-    };
-    assert_eq!(
-        replay_result,
-        Err(SdkError::Conflict("tcp handshake replay detected"))
-    );
-}
-
-#[test]
-fn regression_tampered_tcp_envelope_signature_is_rejected() {
-    // Regression: #822
-    let envelope = match TcpSignedEnvelope::new(
-        did("sender-regression"),
-        did("listener-regression"),
-        1,
-        "state:regression",
-        "expected-body",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("regression envelope build failed: {error}"),
-    };
-    let tampered_payload = format!(
-        "from={}\n\
-to={}\n\
-nonce=1\n\
-state_hash=state:regression\n\
-body=tampered-body-extended\n\
-signer_public_key={}\n\
-signature={}\n",
-        envelope.from.as_str(),
-        envelope.to.as_str(),
-        envelope.signer_public_key,
-        envelope.signature
-    );
-
-    assert_eq!(
-        TcpSignedEnvelope::parse_wire_payload(tampered_payload.as_str()),
-        Err(SdkError::InvalidInput {
-            field: "signature",
-            reason: "failed cryptographic envelope verification",
-        })
-    );
-}
-
-#[test]
-fn regression_forged_handshake_frame_is_rejected() {
-    // Regression: #823
-    let addr = free_addr();
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-
-    let envelope = match TcpSignedEnvelope::new(
-        did("sender-forged"),
-        did("listener-forged"),
-        5,
-        "state:forged",
-        "forged-handshake-frame",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope build failed: {error}"),
-    };
-
-    let forged_payload = format!(
-        "frame=handshake\n\
-version=1\n\
-profile=secp256k1:baseline-v2\n\
-from={}\n\
-to={}\n\
-nonce={}\n\
-signer_public_key={}\n\
-signature=sig:secp256k1:baseline-v2:0:00\n\n{}",
-        envelope.from,
-        envelope.to,
-        envelope.nonce,
-        envelope.signer_public_key,
-        envelope.to_wire_payload()
-    );
-
-    let listener_thread = thread::spawn(move || listener_adapter.listen_once());
-    thread::sleep(Duration::from_millis(30));
-    send_raw_payload(addr.as_str(), forged_payload.as_str());
-
-    let listener_result = match listener_thread.join() {
-        Ok(result) => result,
-        Err(_) => panic!("listener thread panicked"),
-    };
-    assert_eq!(
-        listener_result,
-        Err(SdkError::InvalidInput {
-            field: "handshake.signature",
-            reason: "does not match envelope signature",
-        })
-    );
-}
-
-#[test]
-fn performance_tcp_adapter_local_relay_contract_stays_within_budget() {
-    let addr = free_addr();
-    let listener_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("listener config failed: {error}"),
-    };
-    let sender_config = match TcpTransportConfig::new(addr.as_str()) {
-        Ok(value) => value,
-        Err(error) => panic!("sender config failed: {error}"),
-    };
-
-    let listener_adapter = TcpTransportAdapter::new(listener_config);
-    let sender_adapter = TcpTransportAdapter::new(sender_config);
-
-    let envelope = match TcpSignedEnvelope::new(
-        did("sender-perf"),
-        did("listener-perf"),
-        1,
-        "state:perf",
-        "perf-envelope",
-        TEST_TCP_SIGNING_PRIVATE_KEY_HEX,
-    ) {
-        Ok(value) => value,
-        Err(error) => panic!("envelope build failed: {error}"),
-    };
-
-    let started = Instant::now();
-    let listener_thread = thread::spawn(move || listener_adapter.listen_once());
-    thread::sleep(Duration::from_millis(30));
-
-    if let Err(error) = sender_adapter.send(&envelope) {
-        panic!("sender adapter failed: {error}");
-    }
-
-    let received = match listener_thread.join() {
-        Ok(result) => match result {
-            Ok(value) => value,
-            Err(error) => panic!("listener failed: {error}"),
-        },
-        Err(_) => panic!("listener thread panicked"),
-    };
-    assert_eq!(received.envelope, envelope);
-
-    let elapsed_millis = started.elapsed().as_millis();
-    assert!(
-        elapsed_millis < 500,
-        "tcp adapter relay contract lane exceeded budget: {elapsed_millis}ms"
-    );
-}
+#[path = "tcp_transport_adapter/envelope_validation_contract_tests.rs"]
+mod envelope_validation_contract_tests;
+#[path = "tcp_transport_adapter/relay_flow_contract_tests.rs"]
+mod relay_flow_contract_tests;
+#[path = "tcp_transport_adapter/replay_reconnect_contract_tests.rs"]
+mod replay_reconnect_contract_tests;
+#[path = "tcp_transport_adapter/tamper_rejection_contract_tests.rs"]
+mod tamper_rejection_contract_tests;
+#[path = "tcp_transport_adapter/performance_contract_tests.rs"]
+mod performance_contract_tests;
+#[path = "tcp_transport_adapter/support.rs"]
+mod support;
